db config

+++ src/Controller/user.controller.js
@@ -3,14 +3,20 @@ const user = require('../Model/user.model')
 const deleteUser = (req, res, next) => {
     try {
         const userId = req.params.id
-        user.deleteUser(userId, (err, deleteResponse) => {
-            if (err) {
-                next(err)
-            }
-            else {
-                res.status(200).send("User Deleted")
-            }
-        })
+        if(userId){
+
+            user.deleteUser(userId, (err, deleteResponse) => {
+                if (err) {
+                    next(err)
+                }
+                else {
+                    res.status(200).send("User Deleted")
+                }
+            })
+        }else{
+        res.status(404).send({message:"Missing User ID"})
+
+        }
     } catch (error) {
         next(error)
     }
@@ -19,14 +25,20 @@ const deleteUser = (req, res, next) => {
 const getUser = (req, res, next) => {
     try {
         const userId = req.params.id
-        user.getUser(userId, (err, userResponse) => {
-            if (err) {
-                next(err)
-            }
-            else {
-                res.status(200).send(userResponse)
-            }
-        })
+        if(userId){
+
+            user.getUser(userId, (err, userResponse) => {
+                if (err) {
+                    next(err)
+                }
+                else {
+                    res.status(200).send(userResponse)
+                }
+            })
+        }else{
+        res.status(404).send({message:"Missing User ID"})
+
+        }
     } catch (error) {
         next(error)
     }
@@ -36,14 +48,20 @@ const updateUser = (req, res, next) => {
     try {
         const userId = req.params.id
         const data = req.body
-        user.updateUser(data, userId, (err, userResponse) => {
-            if (err) {
-                next(err)
-            }
-            else {
-                res.status(200).send("User Updated Successfully")
-            }
-        })
+        if(userId){
+
+            user.updateUser(data, userId, (err, userResponse) => {
+                if (err) {
+                    next(err)
+                }
+                else {
+                    res.status(200).send("User Updated Successfully")
+                }
+            })
+        }else{
+        res.status(404).send({message:"Missing User ID"})
+
+        }
     } catch (error) {
         next(error)
     }
